fixes pipelines

+++ src/main/java/com/dashboard/dataTransferObject/revenue/RevenueRead.java
@@ -1,10 +1,11 @@
 package com.dashboard.dataTransferObject.revenue;
 
 import lombok.Data;
+import java.math.BigDecimal;
 
 @Data
 public class RevenueRead {
     private String id;
     private String month;
-    private Double revenue;
+    private BigDecimal revenue;
 }
