+++ src/main/java/com/dashboard/model/entities/Invoice.java
@@ -6,6 +6,7 @@ import org.bson.types.ObjectId;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.DBRef;
 import org.springframework.data.mongodb.core.mapping.Document;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -15,7 +16,7 @@ public class Invoice {
     private ObjectId _id;
     @DBRef
     private Customer customer;
-    private Double amount;
+    private BigDecimal amount;
     private LocalDate date;
     private String status;
     private Audit audit;
